Check ViewController isn't creating retain cycles
Passing the function pointers might look like we're causing implicit
self-references, but as this test confirms, passing the function pointer
only causes an implicit reference to the object the method is on. To
confirm the test catches retain cycles, one can add an `update` method
on the viewController and pass that pointer to `onNext` instead of
`listView.update`.