Replace generic diff, with very context-specific diff
It appears that despite claims that "it almost certainly won't be the
bottleneck in your UI code", the Dwifft diffing algorithm doesn't seem
to scale to hundreds of items.

For the time being, let's make assumptions we know to be currently true
regarding how the data can change from update to update and do a simple,
constant-speed calculation of the diff.